Split svm_transaction.py out of svm.py; reuse canonical is_native_token
- Move send_solana_transaction + confirm_solana_signature into
  svm_transaction.py so the fund-moving surface is isolated from the
  balance/ATA read helpers (svm.py was doing too much).
- Drop the bespoke _NATIVE_SOL_ALIASES set + is_native_sol; native SOL is
  already the sentinel the backend emits, covered by the canonical
  is_native_token in tokens.py. Reuse it.
- Repoint tests; all 36 Solana tests pass.
